+++ newdocs/src/changelog.md
@@ -5,7 +5,8 @@
 <!--- pyml disable-next-line no-duplicate-heading-->
 ### Added
 
-- None
+- [Issue 1212](https://github.com/jackdewinter/pymarkdown/issues/1212)
+    - added cases to Md031 for SetExt
 
 <!--- pyml disable-next-line no-duplicate-heading-->
 ### Fixed